publisher/pyproject.toml
[project]
name = "sbts-publisher"
version = "0.1.0"
authors = [
    { name = "Faiz Surani", email = "faiz@faizsurani.com" },
]
description = "A library for publishing SBTS ballots as PDFs"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
]
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.114.2",
    "pydantic>=2.9.1",
    "gunicorn>=23.0.0",
    "google-cloud-storage>=3.0.0",
    "python-dotenv>=1.0.1",
]

[tool.uv]
dev-dependencies = [
    "ipdb>=0.13.13",
    "pre-commit>=3.8.0",
    "ruff>=0.6.4",
]

[project.urls]
homepage = "https://github.com/reglab/rrc-pipeline"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["publisher"]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
extend-select = ["B", "E", "I", "N", "PTH", "COM", "C4", "UP"]
extend-ignore = ["COM812", "E712", "E501", "B027"]
